+++ src/components/GraphMobile/GraphMobile.module.css
@@ -1,0 +1,43 @@
+.barContainer {
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
+
+.mobileLabel {
+  font-size: 10px;
+  font-family: Roboto;
+  fill: #52555f;
+  margin: 10px 0;
+}
+
+.sections {
+  height: 403px;
+  background-color: #ffffff;
+  box-shadow: 0px 10px 60px rgba(170, 178, 197, 0.2);
+  border-radius: 30px;
+  margin: 0 auto 0;
+  padding: 20px 26px 30px;
+  text-align: center; 
+}
+
+@media (max-width: 771px) {
+  .sections {
+    width: 100%;
+    height: 580px;
+    box-shadow: none;
+  } 
+} 
+
+@media screen and (min-width: 772px) and (max-width: 1279px) {
+  .sections {
+    max-width: 666px;
+  }
+}
+
+@media (min-width: 1280px) {
+  .sections {
+    width: 1060px;
+  }
+}
+
